requirements.txt
requests==2.31.0
jira==3.5.0
python-dotenv==1.0.0
pydantic==2.5.0
GitPython==3.1.40
openai==1.3.0
anthropic==0.7.0
python-gitlab==4.1.0
